+++ build.gradle
@@ -134,6 +134,12 @@ dependencies {
 
     implementation 'com.github.javafaker:javafaker:1.0.2'
 
+
+    //Prometheus
+    implementation "io.micrometer:micrometer-core:1.13.1"
+    implementation "io.micrometer:micrometer-registry-prometheus:1.13.1"
+    implementation "io.micrometer:micrometer-jakarta9:1.13.1"
+
 }
 
 test {
