--- src/main/java/woojooin/planit/global/dto/FCMTokenDto.java
@@ -1,10 +1,0 @@
-package woojooin.planit.global.dto;
-
-import lombok.Data;
-
-@Data
-public class FCMTokenDto {
-    private Long memberId;
-    private String token;
-    private String deviceType; // "web", "android", "ios"
-}
